requirements.txt
# LANGCHAIN 
langchain 
langchain-community 
langchain-google-genai 
langchain-core 
langchain-pinecone

# API AND FRONTEND
fastapi==0.104.1
pydantic==2.5.0
uvicorn 
streamlit 
python-dotenv

# EMBEDDINGS AND VECTOR STORE 
sentence-transformers
pinecone-client
requests
python-multipart
pypdf
